fix: compulab: imx8mm: Address compilation issues with gcc-14.2.0-nolibc

+++ board/compulab/plat/imx8mm/ddr/ddr.c
@@ -20,8 +20,8 @@
 /* Forward declarations */
 u32 cl_eeprom_get_ddrinfo(void);
 u32 cl_eeprom_set_ddrinfo(u32 ddrinfo);
-u32 cl_eeprom_get_subind(void);
-u32 cl_eeprom_set_subind(u32 subind);
+u8 cl_eeprom_get_subind(void);
+u8 cl_eeprom_set_subind(u8 subind);
 
 #ifdef CONFIG_SPL_REPORT_FAKE_MEMSIZE
 u32 cl_eeprom_get_osize(void);
